pyproject.toml
[tool.poetry]
name = "wowguildexport"
version = "0.1.0"
description = "An utils to use the wow api"
authors = ["Armand \"Skengrek\" Massé <armand.masse69@gmail.com>"]
license = "GNU GENERAL PUBLIC LICENSE Version 3"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.4"
requests = "^2.32.3"
apache-airflow = "2.10.1"
python-dotenv = "^1.0.1"
alembic = "^1.13.3"
sqlalchemy = "1.4.54"
psycopg2 = "^2.9.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
